Applied patch from EXLBR-16.
This new class provides a bridge between Logger and OutputStream objects.
It's useful if you have a library that uses an OutputStream or a Writer
destination for output, and you want this to go to a Logger object.

Added test case to validate, and rolled junit version number to be the latest.



git-svn-id: https://svn.apache.org/repos/asf/excalibur/trunk@109818 13f79535-47bb-0310-9956-ffa450edef68